suggested improvements
